+++ Simulation/src/ste/stl_extensions/platform_dependant/thread_affinity.hpp
@@ -28,9 +28,9 @@ bool inline thread_set_affinity(std::thread *thread, const std::bitset<N> &mask)
 
 	DWORD m = 0;
 	if (sizeof(DWORD) == sizeof(decltype(mask.to_ulong())))
-		m = mask.to_ulong();
+		m = static_cast<DWORD>(mask.to_ulong());
 	else if (sizeof(DWORD) == sizeof(decltype(mask.to_ullong())))
-		m = mask.to_ullong();
+		m = static_cast<DWORD>(mask.to_ullong());
 	else
 		assert(false);
